wcore: don't parse/validate the exact major/minor version
Keeping track what is and isn't the correct version at any given time
sounds unfeasible.

On one hand currently we allow any random value as minor for ES2 and
ES3, whist for GL we allow any values as long as they are >= 1.0

If we are to keep track what versions are valid, this is going to cause
problems as new versions of the specs get released.

Let's do the sane thing and remove all the fine-grained validation from
waffle.

v2 (Suggestions from Chad):
 - Keep < 1.0 check to prevent WAFFLE_DONT_CARE as context version.
 - Update the unit tests alongside this intended functionality change.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>