tests/gl_basic: call waffle_error_get_code() less often
There is no need to (potentially) call the function twice in a roll, as
there are no waffle calls inbetween that can trigger a change.

Use a switch (as opposed to if-else 'spaghetti'), which has the
benefit of slightly improved readability.

While we're here, fix the copy/paste in the related comment.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>